Add initial debianizazion
The package is a little wrapper around the k8s GPU plugin:
https://github.com/RadeonOpenCompute/k8s-device-plugin

Upstream suggests to deploy the go binary in a k8s daemonset
with priviledged security settings, plus running as root
(due to the need to accessing dirs like /var/lib/kubelet).
Since we don't really want to allow such access in k8s and
since the go binary can run as simple daemon on "bare metal",
we decided to wrap it in a simple Debian package.

Bug: T333009
Change-Id: I1a8dee05f3f0b788267d11888f4670e62787d386